Upped rev. and changed handling of persistent jack connections.
Now attempts to reconnect arServer ports to client ports that are
announced as available. Previous approach tried to reconnect everything
when ANY new client port was announced, which caused jackd to crash
after a few days or running.

And, after a week of testing, taking the code revision to 4.0.1beta!