Finished UI playlist portion of item editing, and added the UI stash.
Client UI playlist item editing finished.

Client UI stash finished.

Added arServer support for file URL's with prefix/path format via "//"
delimiter in the path between the prfix and the relative portion of the
file path.  This allows arServer to find file URL based on relative
paths and prefixes, just like it does for database items, and was a
needed feature to play files from the temporary media directory.

Added a feature to the libaray management UI panel and associated back
end to allow the option of adding audio files found while crawling the
file system, along with the existing feature of updating found file that
are already in the library but were marked as missing.  This is a useful
tool to in the process of totally rebuilding a library from scratch
using an existing media file library.